Extract shared hit-target audit helpers
